requirements.txt
librosa
matplotlib
numpy
Pillow
tensorflow[and-cuda]
seaborn
gradio